split model adapter
