+++ evalscope/benchmarks/competition_math/__init__.py
@@ -1,6 +1,1 @@
 # Copyright (c) Alibaba, Inc. and its affiliates.
-
-from evalscope.benchmarks.competition_math.competition_math_adapter import DATASET_ID, SUBSET_LIST
-from evalscope.benchmarks.competition_math.competition_math_adapter import CompetitionMathAdapter
-from evalscope.benchmarks.competition_math.competition_math_adapter import CompetitionMathAdapter as DataAdapterClass
-from evalscope.models.model_adapter import ChatGenerationModelAdapter as ModelAdapterClass  # noqa
